Found the problem.
The problem is that the interpolation function fails to extrapolate values for low preassure i.e. eos(4.6650844046460156e-08) for middle guives an error.

Next what I will do is test the issue further and try and find the issue. For now I think that an if could solve this. If there is an issue rho = 0. Lets see where this goes.