Finishd the creation and ploting of Dark matter stars
Moved the creation of data to a function that can both solve neutron stars and dark matter stars.

Changed the get_inpunts functon so the DEBUG data is the parameter you enter. I find this easier than changing the definition of the function.

For the same reason: Not wanting to change functions. I have kept the ploting part outside a function. I know its not the most elegant but its a part of the code that involves a lot of tinkering.