Made a plot to compare the dm mass
I do not like it.

I think, the way the analysis has been set up is not ideal. Instead of fixing a central pressure, in order to make the analysis more comprehensible, we have to fix a total mass.

The analysis would then be that of a star of fixed mass consisting of  baryonic matter and dark matter bound together. The parameter that varies is the mass of the dark matter particles and the graph would show how this affects the other free parameters.

The only issue, I need more code