Finished the stiffnes test
For both dm_masses the trend is the same. As I increase lambda the baryonic radius decreases. But its more noticible for the core case.

Due to this, I'm going to tho a plot that shows how the radius as a function of lambda. So I can compare both masses.

Also something interesting happens with the baryonic central pressure for the halo case.