Started the process for relativistic eos
In this scrip i'll try and create a function that opperates similar to the read_eos function in DANTE.py

This function, due to the further liberty of dark matter masses will need to check if the excel file containing the eos exists, and if that is not the case, create a new one.

For small values of P, i'll use the taylor expansion. Dont really know  why there is an issue but its something.