Finished the graph. Now I need to create a function that finds the alpha for which lambda is max
Finished the graph that  shows how as we increase alpha (for high values) lambda has a maximum.

Now I need to create a function that finds the value for alpha for which lambda is max.

One optional function would be to calculate the avl curve. This one can be done using parallel processing so it can be easily optimized.